Set priority class on Rook 1.0
If rook_priority_class is set then it will be added to Rook Deployments
and DaemonSets on Rook 1.0 only since Rook 1.4 supports configuration of
priority classes directly.
The operator will gradually add priority class to existing objects and
the webhook will ensure that all new objects begin with the priority.