feat: migration ready endpoint returns the number of nodes
to enable the join.sh script to determine the appropriate timing for
approving a storage migration, it requires information about the
current number of nodes in the cluster. By comparing this value with
its own expected count, the script can initiate a migration if they are
equal. this pull request introduces a modification to the migration
ready endpoint, enabling it to respond with the current number of nodes
in the cluster.