feat: storage migration ready endpoint returns the number of nodes (#137)
* feat: migration ready endpoint returns the number of nodes

to enable the join.sh script to determine the appropriate timing for
approving a storage migration, it requires information about the
current number of nodes in the cluster. By comparing this value with
its own expected count, the script can initiate a migration if they are
equal. this pull request introduces a modification to the migration
ready endpoint, enabling it to respond with the current number of nodes
in the cluster.

* Fix spacing in IsMigrationReady() comment

Co-authored-by: Andrew Lavery <laverya@umich.edu>

* Add required number of nodes to migration /ready response

* Update write json marshaling error

Co-authored-by: Ethan Mosbaugh <ethan@replicated.com>

---------

Co-authored-by: Rafael Polanco <6497491+rrpolanco@users.noreply.github.com>
Co-authored-by: Andrew Lavery <laverya@umich.edu>
Co-authored-by: Ethan Mosbaugh <ethan@replicated.com>